+++ src/lib/discovery/bootstrap.ts
@@ -1,4 +1,5 @@
 import debug from 'debug';
+import { Multiaddr } from 'multiaddr';
 
 import { DnsNodeDiscovery } from './dns';
 
@@ -6,7 +7,9 @@ import { getNodesFromHostedJson, getPseudoRandomSubset } from './index';
 
 const dbg = debug('waku:discovery:bootstrap');
 
-const DefaultMaxPeers = 1;
+export const DefaultMaxPeers = 1;
+
+export type BootstrapFn = () => Promise<Multiaddr[]>;
 
 /**
  * Setup discovery method used to bootstrap.
@@ -17,11 +20,11 @@ export interface BootstrapOptions {
   /**
    * The maximum of peers to connect to as part of the bootstrap process.
    *
-   * @default 1
+   * @default {{DefaultMaxPeers}}
    */
   maxPeers?: number;
   /**
-   * Use the default discovery method.
+   * Use the default discovery method. Overrides all other options but `maxPeers`
    *
    * The default discovery method is likely to change overtime as new discovery
    * methods are implemented.
@@ -36,7 +39,7 @@ export interface BootstrapOptions {
   /**
    * Getter that retrieve multiaddrs of peers to connect to.
    */
-  getPeers?: () => Promise<string[]>;
+  getPeers?: () => Promise<string[] | Multiaddr[]>;
   /**
    * An EIP-1459 ENR Tree URL. For example:
    * "enrtree://AOFTICU2XWDULNLZGRMQS4RIZPAZEHYMV4FYHAPW563HNRAOERP7C@test.nodes.vac.dev"
@@ -47,69 +50,54 @@ export interface BootstrapOptions {
 /**
  * Parse the bootstrap options and returns an async function that returns node addresses upon invocation.
  */
-export function parseBootstrap(
-  options:
-    | BootstrapOptions
-    | boolean
-    | string[]
-    | (() => string[] | Promise<string[]>)
-): undefined | (() => Promise<string[]>) {
-  if (
-    Object.keys(options).includes('default') ||
-    Object.keys(options).includes('maxPeers') ||
-    Object.keys(options).includes('peers') ||
-    Object.keys(options).includes('getPeers') ||
-    Object.keys(options).includes('enrUrl')
-  ) {
-    const opts = options as unknown as BootstrapOptions;
-    const maxPeers = opts.maxPeers || DefaultMaxPeers;
+export function parseBootstrap(opts: BootstrapOptions): BootstrapFn {
+  const maxPeers = opts.maxPeers ?? DefaultMaxPeers;
 
-    if (opts.default) {
-      dbg('Bootstrap: Use hosted list of peers.');
+  if (opts.default) {
+    dbg('Bootstrap: Use hosted list of peers.');
 
-      return getNodesFromHostedJson.bind({}, undefined, undefined, maxPeers);
-    } else if (opts.peers !== undefined && opts.peers.length > 0) {
-      dbg('Bootstrap: Use provided list of peers.');
+    return getNodesFromHostedJson.bind({}, undefined, undefined, maxPeers);
+  } else if (opts.peers !== undefined && opts.peers.length > 0) {
+    dbg('Bootstrap: Use provided list of peers.');
 
-      const allPeers: string[] = opts.peers;
-      return (): Promise<string[]> => {
-        const peers = getPseudoRandomSubset(allPeers, maxPeers);
-        return Promise.resolve(peers);
-      };
-    } else if (typeof opts.getPeers === 'function') {
-      dbg('Bootstrap: Use provided getPeers function.');
-      const getPeers = opts.getPeers;
-
-      return async (): Promise<string[]> => {
-        const allPeers = await getPeers();
-        return getPseudoRandomSubset(allPeers, maxPeers);
-      };
-    } else if (opts.enrUrl) {
-      const enrUrl = opts.enrUrl;
-      dbg('Bootstrap: Use provided EIP-1459 ENR Tree URL.');
-
-      const dns = DnsNodeDiscovery.dnsOverHttp();
-      // TODO: The `toString` is incorrect.
-      return (): Promise<string[]> =>
-        dns
-          .getPeers(maxPeers, [enrUrl])
-          .then((peers) => peers.map((peer) => peer.toString()));
-    }
-  } else {
-    dbg(
-      'WARN: This bootstrap method will be deprecated, use `BootstrapOptions` instead'
+    const allPeers: Multiaddr[] = opts.peers.map(
+      (node: string) => new Multiaddr(node)
     );
-    if (options === true) {
-      return getNodesFromHostedJson;
-    } else if (Array.isArray(options)) {
-      return (): Promise<string[]> => {
-        return Promise.resolve(options as string[]);
-      };
-    } else if (typeof options === 'function') {
-      return async (): Promise<string[]> => {
-        return options();
-      };
-    }
+    const peers = getPseudoRandomSubset(allPeers, maxPeers);
+    return (): Promise<Multiaddr[]> => Promise.resolve(peers);
+  } else if (typeof opts.getPeers === 'function') {
+    dbg('Bootstrap: Use provided getPeers function.');
+    const getPeers = opts.getPeers;
+
+    return async (): Promise<Multiaddr[]> => {
+      const allPeers = await getPeers();
+      return getPseudoRandomSubset<string | Multiaddr>(allPeers, maxPeers).map(
+        (node) => new Multiaddr(node)
+      );
+    };
+  } else if (opts.enrUrl) {
+    const enrUrl = opts.enrUrl;
+    dbg('Bootstrap: Use provided EIP-1459 ENR Tree URL.');
+
+    const dns = DnsNodeDiscovery.dnsOverHttp();
+
+    return async (): Promise<Multiaddr[]> => {
+      const enrs = await dns.getPeers(maxPeers, [enrUrl]);
+      const addresses: Multiaddr[] = [];
+      enrs.forEach((enr) => {
+        if (!enr.multiaddrs) return;
+
+        enr.multiaddrs.forEach((ma: Multiaddr) => {
+          // Only return secure websocket addresses
+          if (ma.protoNames().includes('wss')) {
+            addresses.push(ma);
+          }
+        });
+      });
+      return addresses;
+    };
+  } else {
+    dbg('No bootstrap method specified, no peer will be returned');
+    return (): Promise<Multiaddr[]> => Promise.resolve([]);
   }
-  return;
 }